backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-ai-slim[openai]
devtools
